user-actor-my-pets - move registration-handler into lib/handlers
