+++ src/lib/handlers/registration-handler.ts
